Consolidate the bag replicator
