+++ .sbt_metadata/bag_replicator.json
@@ -1,1 +1,1 @@
-{"id" : "bag_replicator", "folder" : "bag_replicator", "dependencyIds" : ["bag_replicator_common"]}
+{"id" : "bag_replicator", "folder" : "bag_replicator", "dependencyIds" : ["common"]}
