+++ bag_replicator/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator/config/ReplicatorDestinationConfig.scala
