+++ bag_replicator/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator/services/UnpackedBagService.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.bagreplicator.unpack_to_archive.services
+package uk.ac.wellcome.platform.archive.bagreplicator.services
 
 import java.io.InputStream
