+++ bag_replicator/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator/BagReplicatorFeatureTest.scala
@@ -1,10 +1,12 @@
-package uk.ac.wellcome.platform.archive.bagreplicator.unpack_to_archive
+package uk.ac.wellcome.platform.archive.bagreplicator
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.BagReplicatorFixtures
-import uk.ac.wellcome.platform.archive.bagreplicator.unpack_to_archive.fixtures.WorkerServiceFixture
+import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.{
+  BagReplicatorFixtures,
+  WorkerServiceFixture
+}
 import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
 import uk.ac.wellcome.platform.archive.common.generators.BagRequestGenerators
 import uk.ac.wellcome.platform.archive.common.models.BagRequest
