+++ bag_replicator/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator/services/BagReplicatorWorkerServiceTest.scala
@@ -1,13 +1,15 @@
-package uk.ac.wellcome.platform.archive.bagreplicator.archive_to_access.services
+package uk.ac.wellcome.platform.archive.bagreplicator.services
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.platform.archive.bagreplicator.archive_to_access.fixtures.WorkerServiceFixture
-import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.BagReplicatorFixtures
+import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.{
+  BagReplicatorFixtures,
+  WorkerServiceFixture
+}
 import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
 import uk.ac.wellcome.platform.archive.common.generators.BagRequestGenerators
-import uk.ac.wellcome.platform.archive.common.models.ReplicationResult
+import uk.ac.wellcome.platform.archive.common.models.BagRequest
 import uk.ac.wellcome.platform.archive.common.models.bagit.{
   BagLocation,
   BagPath
@@ -26,35 +28,37 @@ class BagReplicatorWorkerServiceTest
     with WorkerServiceFixture {
 
   it("replicates a bag successfully and updates both topics") {
-    withLocalS3Bucket { bucket =>
-      withLocalSnsTopic { progressTopic =>
-        withLocalSnsTopic { outgoingTopic =>
-          withWorkerService(
-            progressTopic = progressTopic,
-            outgoingTopic = outgoingTopic) { service =>
-            withBag(bucket) { srcBagLocation =>
-              val bagRequest = createBagRequestWith(srcBagLocation)
+    withLocalS3Bucket { ingestsBucket =>
+      withLocalS3Bucket { archiveBucket =>
+        val destination = createReplicatorDestinationConfigWith(archiveBucket)
+        withLocalSnsTopic { progressTopic =>
+          withLocalSnsTopic { outgoingTopic =>
+            withWorkerService(
+              progressTopic = progressTopic,
+              outgoingTopic = outgoingTopic,
+              destination = destination) { service =>
+              withBag(ingestsBucket) { srcBagLocation =>
+                val bagRequest = createBagRequestWith(srcBagLocation)
 
-              val future = service.processMessage(bagRequest)
+                val future = service.processMessage(bagRequest)
 
-              whenReady(future) { _ =>
-                val result =
-                  notificationMessage[ReplicationResult](outgoingTopic)
-                result.requestId shouldBe bagRequest.requestId
-                result.srcBagLocation shouldBe bagRequest.bagLocation
+                whenReady(future) { _ =>
+                  val result = notificationMessage[BagRequest](outgoingTopic)
+                  result.requestId shouldBe bagRequest.requestId
 
-                val dstBagLocation = result.dstBagLocation
+                  val dstBagLocation = result.bagLocation
 
-                verifyBagCopied(
-                  src = srcBagLocation,
-                  dst = dstBagLocation
-                )
+                  verifyBagCopied(
+                    src = srcBagLocation,
+                    dst = dstBagLocation
+                  )
 
-                assertTopicReceivesProgressEventUpdate(
-                  bagRequest.requestId,
-                  progressTopic) { events =>
-                  events should have size 1
-                  events.head.description shouldBe "Bag replicated successfully"
+                  assertTopicReceivesProgressEventUpdate(
+                    bagRequest.requestId,
+                    progressTopic) { events =>
+                    events should have size 1
+                    events.head.description shouldBe "Bag successfully copied from ingest location"
+                  }
                 }
               }
             }
@@ -89,7 +93,7 @@ class BagReplicatorWorkerServiceTest
               progressTopic = progressTopic,
               status = Progress.Failed) { events =>
               events should have size 1
-              events.head.description shouldBe s"Failed to replicate bag"
+              events.head.description shouldBe "Failed to copy bag from ingest location"
             }
           }
         }
