--- bag_replicator_common/docker-compose.yml
@@ -1,13 +1,0 @@
-sqs:
-  image: s12v/elasticmq
-  ports:
-    - "9324:9324"
-    - "4789:9324"
-sns:
-  image: wellcome/fake-sns
-  ports:
-    - "9292:9292"
-s3:
-  image: scality/s3server:mem-latest
-  ports:
-    - "33333:8000"
