--- bag_replicator_common/src/main/resources/application.conf
@@ -1,5 +1,0 @@
-aws.sqs.queue.url=${?queue_url}
-aws.progress.sns.topic.arn=${?progress_topic_arn}
-aws.outgoing.sns.topic.arn=${?outgoing_topic_arn}
-bag-replicator.storage.destination.bucket=${?destination_bucket_name}
-metrics.namespace=bag-replicator
