--- bag_replicator_common/src/test/resources/application.conf
@@ -1,3 +1,0 @@
-akka {
-  loglevel = DEBUG
-}
