--- bag_replicator_unpack_to_archive/Dockerfile
@@ -1,5 +1,0 @@
-FROM wellcome/typesafe_config_base:27
-
-ADD target/universal/stage /opt/docker
-
-ENV PROJECT bag_replicator_unpack_to_archive
