--- bag_replicator_unpack_to_archive/src/main/resources/logback.xml
@@ -1,16 +1,0 @@
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-    </encoder>
-  </appender>
-
-  <root level="${log_level:-INFO}">
-    <appender-ref ref="STDOUT" />
-  </root>
-
-  <!-- reduce external logging -->
-  <logger name="org.apache.http" level="ERROR"/>
-  <logger name="io.netty" level="ERROR"/>
-  <logger name="com.amazonaws" level="WARN"/>
-</configuration>
