--- bag_replicator_unpack_to_archive/src/test/resources/application.conf
@@ -1,3 +1,0 @@
-akka {
-  loglevel = DEBUG
-}
